% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging.R
\name{annual_averages}
\alias{annual_averages}
\title{Calculate Annual Averages for a variable}
\usage{
annual_averages(var, file_stem, years, geoid = "zip",
  outdir = "../data/out")
}
\arguments{
\item{var}{variable name containing the value to be averaged}

\item{file_stem}{the stem of all files that contain daily data for a variable. Should include
paths along with the file stem. Assumes that all files have this name followed
by the year of data. For example, temperature data named "tmmx_zip_yyyy.csv" stored in
"data/temperature" would have "data/temperature/tmmx_zip_" as the file stem.}

\item{years}{years to create annual averages for.}

\item{geoid}{variable name containing the unique ID for each geographical area.}

\item{outdir}{Directory to save output in.  Output data is saved there
in a file names "[var]_annual_averages.csv"}
}
\description{
Calculate Annual Averages for a variable
}
