% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet.R
\name{download_gridmet}
\alias{download_gridmet}
\title{Download grid met data and store in the temporary directory}
\usage{
download_gridmet(variable, year, temp_dirname)
}
\arguments{
\item{variable}{Gridmet variable to download}

\item{year}{year of data to download}

\item{temp_dirname}{directory to store the downloaded file in}
}
\description{
Download grid met data and store in the temporary directory
}
