% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet.R
\name{get_gridmet}
\alias{get_gridmet}
\title{Download Gridmet data set, aggregate to zip code or county}
\usage{
get_gridmet(variable, years = 1999:2019, poly, points = NULL,
  temp_dirname = "temp", outdir = "out", cache = T)
}
\arguments{
\item{variable}{name of the variable to download}

\item{years}{vector of years of data to process}

\item{poly}{sf spatial polygoon object}

\item{points}{sf spatial points object}

\item{temp_dirname}{where should downloaded data be stored}

\item{outdir}{directory to store the output data}

\item{cache}{should the downloaded data be stored}
}
\description{
Download Gridmet data set, aggregate to zip code or county
}
