% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographies.R
\name{get_year_zip_shapefiles}
\alias{get_year_zip_shapefiles}
\title{Read in polygon and point shape from the standard
zip code structure}
\usage{
get_year_zip_shapefiles(shape_path, year)
}
\arguments{
\item{shape_path}{path to the directory containing the zip code shape files}

\item{year}{year of zip codes to grab}
}
\value{
a list with the polygon and point shpae files foor a given year
}
\description{
Read in polygon and point shape from the standard
zip code structure
}
\details{
Assumes the folder structure in the zip_for_loop directory
}
