% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet.R
\name{process_year_polygon}
\alias{process_year_polygon}
\title{Create a data frame of daily zip code area weighted estiamtes}
\usage{
process_year_polygon(rast, poly, variable, geoid.in, geoid.out, subdivide)
}
\description{
Create a data frame of daily zip code area weighted estiamtes
}
