% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging.R
\name{seasonal_averages}
\alias{seasonal_averages}
\title{Calculate seasonal averages for a variable}
\usage{
seasonal_averages(var, file_stem, years, geoid = "zip",
  outdir = "../data/out")
}
\arguments{
\item{var}{variable name containing the value to be averaged}

\item{file_stem}{the stem of all files that contain daily data for a variable. Should include
paths along with the file stem. Assumes that all files have this name followed
by the year of data. For example, temperature data named "tmmx_zip_yyyy.csv" stored in
"data/temperature" would have "data/temperature/tmmx_zip_" as the file stem.}

\item{years}{years to create seasonal averages for.}

\item{geoid}{variable name containing the unique ID for each geographical area.}

\item{outdir}{Directory to save output in.  Output data is saved there
in a file names "[var]_seasonal_averages.csv"}
}
\description{
Calculate seasonal averages for a variable
}
\details{
Assumes that for a given year \code{y}, winter runs from
December 1st, y-1 to February 28/29, y and that summer
runs from June 1st, y to August 31st, y. Future functionality
may allow for customization of winter and summer intervals. If data
for a previous year is not available, data from the start of the year, onward is used.
}
