% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet.R
\name{get_gridmet}
\alias{get_gridmet}
\title{Download Gridmet data set, aggregate to zip code or county}
\usage{
get_gridmet(variable, years = 1999:2019, poly, points = NULL,
  geoid.in = "ZIP", geoid.out = "zip", temp_dirname = "temp",
  outdir = "out", geo_name = "zip", subdivide = T, cache = T)
}
\arguments{
\item{variable}{name of the variable to download}

\item{years}{vector of years of data to process}

\item{poly}{sf spatial polygoon object}

\item{points}{sf spatial points object}

\item{geoid.in}{The name of the column in the input shapefile(s) that has the ID of the geographical area}

\item{geoid.out}{What to name the geographic ID in the ooutput data}

\item{temp_dirname}{where should downloaded data be stored}

\item{outdir}{directory to store the output data}

\item{subdivide}{Should downloaded rasters be subdivided for finer area weighting?}

\item{cache}{should the downloaded data be stored}

\item{geoname}{What to name the geogrpahy in the output file}
}
\description{
Download Gridmet data set, aggregate to zip code or county
}
