% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmet.R
\name{get_gridmet}
\alias{get_gridmet}
\title{Download Gridmet data set, aggregate to zip code or county}
\usage{
get_gridmet(variable, years = 1999:2019, geography, geo_type = NULL,
  temp_dirname = "temp", cache = T)
}
\arguments{
\item{variable}{name of the variable to download}

\item{years}{vector of years of data to process}

\item{geography}{what geography to aggregate to, should be zip code or county}

\item{geo_type}{for zip code, should points, polygons, or combined be used.
must be any value in c('polygon','point','combined') Defaults to combined.}

\item{temp_dirname}{where should downloaded data be stored}

\item{cache}{should the downloade data be stored}
}
\description{
Download Gridmet data set, aggregate to zip code or county
}
